Finished transferring board methods to object
All the methods relating to the game board have been transferred to an object with the name Board. This removes many of the unnecessary return statements for the board array, as well as simplifying the code in the main TicTacToeComputer.java class, and making it easier to modify the board functionality through the Board.java Class.